requirements.txt
python-telegram-bot==20.7
fastapi==0.109.2
uvicorn[standard]==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
supabase==2.3.4
httpx==0.26.0
python-dotenv==1.0.1
aiohttp==3.9.3
apscheduler==3.10.4
pytz==2024.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
itsdangerous==2.1.2
starlette==0.36.3
